Remove select component in favor of MUIs select
And fix some select styling issues:

- Set a fixed height for the select to prevent it from increasing its
  height when, e.g., a CircularProgress is used as the content.

- Set a fixed padding-right for the InputBase to make sure the padding
  is the same, with or without an end-adornment, as the InputBase is
  not aware of the Select's IconComponent with the ChevronDown.
